+++ base/int.jl
@@ -93,9 +93,9 @@ inv(x::Integer) = float(one(x)) / float(x)
 (/)(x::BitInteger, y::BitInteger) = float(x) / float(y)
 
 """
-    isodd(x::Integer) -> Bool
+    isodd(x::Number) -> Bool
 
-Return `true` if `x` is odd (that is, not divisible by 2), and `false` otherwise.
+Return `true` if `x` is an odd integer (that is, an integer not divisible by 2), and `false` otherwise.
 
 # Examples
 ```jldoctest
@@ -106,12 +106,13 @@ julia> isodd(10)
 false
 ```
 """
-isodd(n::Integer) = rem(n, 2) != 0
+isodd(n::Number) = isreal(n) && isodd(real(n))
+isodd(n::Real) = isinteger(n) && !iszero(rem(Integer(n), 2))
 
 """
-    iseven(x::Integer) -> Bool
+    iseven(x::Number) -> Bool
 
-Return `true` if `x` is even (that is, divisible by 2), and `false` otherwise.
+Return `true` if `x` is an even integer (that is, an integer divisible by 2), and `false` otherwise.
 
 # Examples
 ```jldoctest
@@ -122,7 +123,8 @@ julia> iseven(10)
 true
 ```
 """
-iseven(n::Integer) = !isodd(n)
+iseven(n::Number) = isreal(n) && iseven(real(n))
+iseven(n::Real) = isinteger(n) && iszero(rem(Integer(n), 2))
 
 signbit(x::Integer) = x < 0
 signbit(x::Unsigned) = false