+++ base/options.jl
@@ -45,6 +45,7 @@ struct JLOptions
     image_file_specified::Int8
     warn_scope::Int8
     image_codegen::Int8
+    rr_detach::Int8
 end
 
 # This runs early in the sysimage != is not defined yet