+++ src/gc-stacks.c
@@ -216,8 +216,12 @@ void sweep_stack_pools(void)
             continue;
         while (1) {
             jl_task_t *t = (jl_task_t*)lst[n];
+            assert(jl_is_task(t));
             if (gc_marked(jl_astaggedvalue(t)->bits.gc)) {
-                n++;
+                if (t->stkbuf == NULL)
+                    ndel++; // jl_release_task_stack called
+                else
+                    n++;
             }
             else {
                 ndel++;
@@ -242,4 +246,31 @@ void sweep_stack_pools(void)
         }
         live_tasks->len -= ndel;
     }
+}
+
+JL_DLLEXPORT jl_array_t *jl_live_tasks(void)
+{
+    jl_ptls_t ptls = jl_get_ptls_states();
+    arraylist_t *live_tasks = &ptls->heap.live_tasks;
+    size_t i, j, l;
+    jl_array_t *a;
+    do {
+        l = live_tasks->len;
+        a = jl_alloc_vec_any(l + 1); // may gc
+    } while (l + 1 < live_tasks->len);
+    l = live_tasks->len;
+    void **lst = live_tasks->items;
+    j = 0;
+    ((void**)jl_array_data(a))[j++] = ptls->root_task;
+    for (i = 0; i < l; i++) {
+        if (((jl_task_t*)lst[i])->stkbuf != NULL)
+            ((void**)jl_array_data(a))[j++] = lst[i];
+    }
+    l = jl_array_len(a);
+    if (j < l) {
+        JL_GC_PUSH1(&a);
+        jl_array_del_end(a, l - j);
+        JL_GC_POP();
+    }
+    return a;
 }